+++ library/springboot/springboot/src/test/kotlin/de/lise/fluxflow/springboot/ioc/testservices/DuplicateService.kt
@@ -1,0 +1,3 @@
+package de.lise.fluxflow.springboot.ioc.testservices
+
+interface DuplicateService