+++ library/springboot/springboot/src/test/kotlin/de/lise/fluxflow/springboot/ioc/testservices/DuplicateServiceImpl2.kt
@@ -1,0 +1,6 @@
+package de.lise.fluxflow.springboot.ioc.testservices
+
+import org.springframework.stereotype.Service
+
+@Service
+class DuplicateServiceImpl2 : DuplicateService